requirements.txt
# Core
pandas==2.2.2
numpy==1.26.4
scipy==1.13.0

# Visualization
matplotlib==3.8.4
seaborn==0.13.2
plotly==5.22.0

# EDA Automation
ydata-profiling==4.8.3
sweetviz==2.3.1

# ML
scikit-learn==1.4.2
xgboost==2.0.3
lightgbm==4.3.0
imbalanced-learn==0.12.2
statsmodels==0.14.2

# Deep Learning (optional, for LSTM notebook)
# tensorflow==2.16.1

# Backend
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
python-dotenv==1.0.1
httpx==0.27.0

# Frontend
dash==2.17.0
dash-bootstrap-components==1.6.0
dash-ag-grid==31.2.0

# Utilities
joblib==1.4.2
PyYAML==6.0.1
pyarrow==16.0.0
tqdm==4.66.4
pytest==8.2.0
pytest-asyncio==0.23.6
requests==2.32.2
